Extract module viewer into its own file
